+++ src/App/Documentation/components/MediaObject/index.js
@@ -8,8 +8,8 @@ const Overview = () => (
         <h2 id="overview">Overview</h2>
         <p>Use the <Property value=".media" /> class on an object to style it as a small container that feature a left- or right-aligned image/icon alongside textual content.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
-            <MediaObjectComponent imgUrl="https://via.placeholder.com/48x48" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
-            <MediaObjectComponent icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
+            <MediaObjectComponent imgUrl="https://via.placeholder.com/48x48" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
+            <MediaObjectComponent icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
         </ComponentPreview>
     </>
 );
@@ -22,13 +22,13 @@ const Sizes = () => (
         </p>
         <h3>Small</h3>
         <ComponentPreview language="html" showCasePanel codeFigure>
-            <MediaObjectComponent size="sm" imgUrl="https://via.placeholder.com/40x40" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
-            <MediaObjectComponent size="sm" icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
+            <MediaObjectComponent size="sm" imgUrl="https://via.placeholder.com/40x40" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
+            <MediaObjectComponent size="sm" icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
         </ComponentPreview>
         <h3>Large</h3>
         <ComponentPreview language="html" showCasePanel codeFigure>
-            <MediaObjectComponent size="lg" imgUrl="https://via.placeholder.com/56x56" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
-            <MediaObjectComponent size="lg" icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
+            <MediaObjectComponent size="lg" imgUrl="https://via.placeholder.com/56x56" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
+            <MediaObjectComponent size="lg" icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
         </ComponentPreview>
     </>
 );
@@ -38,8 +38,8 @@ const MediaPosition = () => (
         <h2 id="media-position">Media position</h2>
         <p>Use the <Property value=".media-right" /> class to align the media content to the right.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
-            <MediaObjectComponent mediaRight imgUrl="https://via.placeholder.com/48x48" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
-            <MediaObjectComponent mediaRight icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@payex.com" />
+            <MediaObjectComponent mediaRight imgUrl="https://via.placeholder.com/48x48" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
+            <MediaObjectComponent mediaRight icon="account_circle" heading="Bob Corlsan" text="bob.corlsan@swedbankpay.com" />
         </ComponentPreview>
     </>
 );