+++ src/App/Documentation/dashboard/Introduction/index.js
@@ -4,10 +4,10 @@ import PrismCode from "react-prism";
 import { ComponentPreview, Property, DocContainer } from "#";
 
 const BASENAME = process.env.basename;
-const scriptUrl = `https://design.payex.com${BASENAME}scripts/px-script.js`;
-const dashboardScriptUrl = `https://design.payex.com${BASENAME}scripts/px.dashboard.js`;
-const styleUrl = `https://design.payex.com${BASENAME}styles/px.css`;
-const dashboardStyleUrl = `https://design.payex.com${BASENAME}styles/px.dashboard.css`;
+const scriptUrl = `https://design.swedbankpay.com${BASENAME}scripts/px-script.js`;
+const dashboardScriptUrl = `https://design.swedbankpay.com${BASENAME}scripts/px.dashboard.js`;
+const styleUrl = `https://design.swedbankpay.com${BASENAME}styles/px.css`;
+const dashboardStyleUrl = `https://design.swedbankpay.com${BASENAME}styles/px.dashboard.css`;
 
 const QuickStart = () => (
     <>
@@ -40,7 +40,7 @@ const JavaScript = () => (
 
 const Introduction = () => (
     <DocContainer>
-        <p className="lead">Get started with the PayEx DesignGuide Dashboard theme.</p>
+        <p className="lead">Get started with the Swedbank Pay DesignGuide Dashboard theme.</p>
         <QuickStart />
         {/* <Css /> */}
         <JavaScript />