+++ src/App/Documentation/utils/DocContainer/index.js
@@ -1,13 +1,14 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { DocToc } from "#";
+import { DocToc, DocHeading } from "#";
 
 const DocContainer = ({ docToc, children }) => {
     const DocContent = () => <div className={`doc-body ${docToc ? "col-lg-10" : "col-12"}`}>{children}</div>;
 
     return (
         <div className="doc-container">
+            <DocHeading />
             <div className="row">
                 <DocContent />
                 {docToc ? <DocToc component={DocContent} /> : null}