+++ src/App/Documentation/utils/DocToc/index.js
@@ -8,7 +8,7 @@ const _findHeadings = children => {
 
     if (typeof children.map === "function") {
         children.map(child => {
-            if (child.type.name !== "ExperimentalComponentAlert" && typeof child.type === "function") {
+            if (child.type.name !== "ExperimentalComponentAlert" && typeof child.type === "function" && child.type.displayName !== "withRouter(DocHeading)") {
                 if (child.type().props.children.type === "h2") {
                     headings.push({
                         title: child.type().props.children.props.children,