+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-## [3.1.1] - 2019-06-25
+## [3.2.0] - 2019-07-12
+
+### Added
+
+- Added .page-footer-rights class to use with .page-footer.
+- Added better feedback when initializing expandables/accordions
 
 ### Changed
-- Changed action-link arrow to be material-icons instead of plain text.
+
+- There is no longer a width constraint on the card component. Use grid to set sizing.
+- Changed front page and general documentation styling.
+- Replaced documentation occurences of PayEx with Swedbank Pay.
+
+### Fixed
+
+- Fixed issue with px.sheet.open and px.sheet.close.
+
+### Removed
+
+- Removed box-shadow from steps component.
+- Removed .cards wrapper.